Options are now given by an (harcoded) extractor
