+++ src/symfony/src/Service/DefaultPublicKeyCredentialCreationOptionsExtractor.php
@@ -1,0 +1,111 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Webauthn\Bundle\Service;
+
+use function count;
+use const FILTER_VALIDATE_BOOLEAN;
+use function is_array;
+use RuntimeException;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
+use Symfony\Component\Serializer\SerializerInterface;
+use Symfony\Component\Validator\Validator\ValidatorInterface;
+use Webauthn\AuthenticationExtensions\AuthenticationExtensionsClientInputs;
+use Webauthn\AuthenticatorSelectionCriteria;
+use Webauthn\Bundle\Dto\PublicKeyCredentialCreationOptionsRequest;
+use Webauthn\PublicKeyCredentialCreationOptions;
+use Webauthn\PublicKeyCredentialDescriptor;
+use Webauthn\PublicKeyCredentialSource;
+use Webauthn\PublicKeyCredentialSourceRepository;
+use Webauthn\PublicKeyCredentialUserEntity;
+
+final class DefaultPublicKeyCredentialCreationOptionsExtractor implements PublicKeyCredentialCreationOptionsExtractor
+{
+    public function __construct(
+        private readonly SerializerInterface $serializer,
+        private readonly ValidatorInterface $validator,
+        private readonly PublicKeyCredentialSourceRepository $credentialSourceRepository,
+        private readonly PublicKeyCredentialCreationOptionsFactory $publicKeyCredentialCreationOptionsFactory,
+        private readonly string $profile,
+    ) {
+    }
+
+    public function getFromRequest(
+        Request $request,
+        PublicKeyCredentialUserEntity $userEntity
+    ): PublicKeyCredentialCreationOptions {
+        $request->getContentType() === 'json' || throw new BadRequestHttpException(
+            'Only JSON content type allowed'
+        );
+        $content = $request->getContent();
+
+        $excludedCredentials = $this->getCredentials($userEntity);
+        $creationOptionsRequest = $this->getServerPublicKeyCredentialCreationOptionsRequest($content);
+        $authenticatorSelectionData = $creationOptionsRequest->authenticatorSelection;
+        $authenticatorSelection = null;
+        if (is_array($authenticatorSelectionData)) {
+            $authenticatorSelection = AuthenticatorSelectionCriteria::createFromArray($authenticatorSelectionData);
+        } elseif ($creationOptionsRequest->userVerification !== null || $creationOptionsRequest->residentKey !== null || $creationOptionsRequest->authenticatorAttachment !== null) {
+            $authenticatorSelection = AuthenticatorSelectionCriteria::create()
+                ->setUserVerification(
+                    $creationOptionsRequest->userVerification ?? AuthenticatorSelectionCriteria::USER_VERIFICATION_REQUIREMENT_PREFERRED
+                )
+                ->setAuthenticatorAttachment($creationOptionsRequest->authenticatorAttachment);
+            if ($creationOptionsRequest->residentKey !== null) {
+                $authenticatorSelection->setResidentKey($creationOptionsRequest->residentKey);
+            }
+            if ($creationOptionsRequest->requireResidentKey !== null) {
+                $authenticatorSelection->setRequireResidentKey(
+                    filter_var($creationOptionsRequest->requireResidentKey, FILTER_VALIDATE_BOOLEAN)
+                );
+            }
+        }
+        $extensions = $creationOptionsRequest->extensions;
+        if (is_array($extensions)) {
+            $extensions = AuthenticationExtensionsClientInputs::createFromArray($extensions);
+        }
+
+        return $this->publicKeyCredentialCreationOptionsFactory->create(
+            $this->profile,
+            $userEntity,
+            $excludedCredentials,
+            $authenticatorSelection,
+            $creationOptionsRequest->attestation,
+            $extensions
+        );
+    }
+
+    /**
+     * @return PublicKeyCredentialDescriptor[]
+     */
+    private function getCredentials(PublicKeyCredentialUserEntity $userEntity): array
+    {
+        $credentialSources = $this->credentialSourceRepository->findAllForUserEntity($userEntity);
+
+        return array_map(
+            static fn (PublicKeyCredentialSource $credential): PublicKeyCredentialDescriptor => $credential->getPublicKeyCredentialDescriptor(),
+            $credentialSources
+        );
+    }
+
+    private function getServerPublicKeyCredentialCreationOptionsRequest(
+        string $content
+    ): PublicKeyCredentialCreationOptionsRequest {
+        $data = $this->serializer->deserialize($content, PublicKeyCredentialCreationOptionsRequest::class, 'json');
+        $data instanceof PublicKeyCredentialCreationOptionsRequest || throw new BadRequestHttpException(
+            'Invalid data'
+        );
+        $errors = $this->validator->validate($data);
+        if (count($errors) > 0) {
+            $messages = [];
+            foreach ($errors as $error) {
+                $messages[] = $error->getPropertyPath() . ': ' . $error->getMessage();
+            }
+            throw new RuntimeException(implode("\n", $messages));
+        }
+
+        return $data;
+    }
+}
