+++ src/symfony/src/Service/DefaultPublicKeyCredentialRequestOptionsExtractor.php
@@ -1,0 +1,96 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Webauthn\Bundle\Service;
+
+use function count;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
+use Symfony\Component\Serializer\Normalizer\AbstractObjectNormalizer;
+use Symfony\Component\Serializer\SerializerInterface;
+use Symfony\Component\Validator\Validator\ValidatorInterface;
+use Webauthn\AuthenticationExtensions\AuthenticationExtensionsClientInputs;
+use Webauthn\Bundle\Dto\ServerPublicKeyCredentialRequestOptionsRequest;
+use Webauthn\Bundle\Repository\PublicKeyCredentialUserEntityRepository;
+use Webauthn\PublicKeyCredentialDescriptor;
+use Webauthn\PublicKeyCredentialRequestOptions;
+use Webauthn\PublicKeyCredentialSource;
+use Webauthn\PublicKeyCredentialSourceRepository;
+use Webauthn\PublicKeyCredentialUserEntity;
+
+final class DefaultPublicKeyCredentialRequestOptionsExtractor implements PublicKeyCredentialRequestOptionsExtractor
+{
+    public function __construct(
+        private readonly SerializerInterface $serializer,
+        private readonly ValidatorInterface $validator,
+        private readonly PublicKeyCredentialUserEntityRepository $userEntityRepository,
+        private readonly PublicKeyCredentialSourceRepository $credentialSourceRepository,
+        private readonly PublicKeyCredentialRequestOptionsFactory $publicKeyCredentialRequestOptionsFactory,
+        private readonly string $profile,
+    ) {
+    }
+
+    public function getFromRequest(
+        Request $request,
+        ?PublicKeyCredentialUserEntity &$userEntity = null
+    ): PublicKeyCredentialRequestOptions {
+        $request->getContentType() === 'json' || throw new BadRequestHttpException(
+            'Only JSON content type allowed'
+        );
+        $content = $request->getContent();
+        $creationOptionsRequest = $this->getServerPublicKeyCredentialRequestOptionsRequest($content);
+        $extensions = $creationOptionsRequest->extensions !== null ? AuthenticationExtensionsClientInputs::createFromArray(
+            $creationOptionsRequest->extensions
+        ) : null;
+        $userEntity = $creationOptionsRequest->username === null ? null : $this->userEntityRepository->findOneByUsername(
+            $creationOptionsRequest->username
+        );
+        $allowedCredentials = $userEntity === null ? [] : $this->getCredentials($userEntity);
+        return $this->publicKeyCredentialRequestOptionsFactory->create(
+            $this->profile,
+            $allowedCredentials,
+            $creationOptionsRequest->userVerification,
+            $extensions
+        );
+    }
+
+    /**
+     * @return PublicKeyCredentialDescriptor[]
+     */
+    private function getCredentials(PublicKeyCredentialUserEntity $userEntity): array
+    {
+        $credentialSources = $this->credentialSourceRepository->findAllForUserEntity($userEntity);
+
+        return array_map(
+            static fn (PublicKeyCredentialSource $credential): PublicKeyCredentialDescriptor => $credential->getPublicKeyCredentialDescriptor(),
+            $credentialSources
+        );
+    }
+
+    private function getServerPublicKeyCredentialRequestOptionsRequest(
+        string $content
+    ): ServerPublicKeyCredentialRequestOptionsRequest {
+        $data = $this->serializer->deserialize(
+            $content,
+            ServerPublicKeyCredentialRequestOptionsRequest::class,
+            'json',
+            [
+                AbstractObjectNormalizer::DISABLE_TYPE_ENFORCEMENT => true,
+            ]
+        );
+        $data instanceof ServerPublicKeyCredentialRequestOptionsRequest || throw new BadRequestHttpException(
+            'Invalid data'
+        );
+        $errors = $this->validator->validate($data);
+        if (count($errors) > 0) {
+            $messages = [];
+            foreach ($errors as $error) {
+                $messages[] = $error->getPropertyPath() . ': ' . $error->getMessage();
+            }
+            throw new BadRequestHttpException(implode("\n", $messages));
+        }
+
+        return $data;
+    }
+}
