+++ src/symfony/src/Service/PublicKeyCredentialRequestOptionsExtractor.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Webauthn\Bundle\Service;
+
+use Symfony\Component\HttpFoundation\Request;
+use Webauthn\PublicKeyCredentialRequestOptions;
+use Webauthn\PublicKeyCredentialUserEntity;
+
+interface PublicKeyCredentialRequestOptionsExtractor
+{
+    public function getFromRequest(
+        Request $request,
+        ?PublicKeyCredentialUserEntity &$userEntity = null
+    ): PublicKeyCredentialRequestOptions;
+}
